+++ src/main/java/co/edu/unicauca/laborDocente/api/dto/AtributoDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class AtributoDTO {
-    private Long oideavAtributo;
+    private Integer oideavAtributo;
     private String nombre;
     private String tipoDato;
 }
